+++ src/LoveServiceProvider.php
@@ -21,18 +21,13 @@ use Cog\Laravel\Love\Console\Commands\SetupReactable;
 use Cog\Laravel\Love\Console\Commands\SetupReacterable;
 use Cog\Laravel\Love\Console\Commands\UpgradeV5ToV6;
 use Cog\Laravel\Love\Console\Commands\UpgradeV7ToV8;
-use Cog\Laravel\Love\Reactant\Listeners\DecrementAggregates;
-use Cog\Laravel\Love\Reactant\Listeners\IncrementAggregates;
 use Cog\Laravel\Love\Reactant\ReactionCounter\Models\ReactionCounter;
 use Cog\Laravel\Love\Reactant\ReactionCounter\Observers\ReactionCounterObserver;
 use Cog\Laravel\Love\Reactant\ReactionTotal\Models\ReactionTotal;
 use Cog\Laravel\Love\Reactant\ReactionTotal\Observers\ReactionTotalObserver;
-use Cog\Laravel\Love\Reaction\Events\ReactionHasBeenAdded;
-use Cog\Laravel\Love\Reaction\Events\ReactionHasBeenRemoved;
 use Cog\Laravel\Love\Reaction\Models\Reaction;
 use Cog\Laravel\Love\Reaction\Observers\ReactionObserver;
 use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\Event;
 use Illuminate\Support\ServiceProvider;
 
 final class LoveServiceProvider extends ServiceProvider
@@ -58,7 +53,6 @@ final class LoveServiceProvider extends ServiceProvider
         $this->registerObservers();
         $this->registerPublishes();
         $this->registerMigrations();
-        $this->registerListeners();
     }
 
     /**
@@ -132,17 +126,6 @@ final class LoveServiceProvider extends ServiceProvider
         if ($this->app->runningInConsole() && $this->shouldLoadDefaultMigrations()) {
             $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
         }
-    }
-
-    /**
-     * Register the Love event listeners.
-     *
-     * @return void
-     */
-    private function registerListeners(): void
-    {
-        Event::listen(ReactionHasBeenAdded::class, IncrementAggregates::class);
-        Event::listen(ReactionHasBeenRemoved::class, DecrementAggregates::class);
     }
 
     /**